+++ src/main/java/com/increff/employee/model/SalesData.java
@@ -7,7 +7,7 @@ public class SalesData {
     private String brandName;
     private String brandCategory;
     private Integer quantity;
-    private Integer revenue;
+    private Float revenue;
     private String barcode;
 
 }
